Change folder structure for drive providers
