+++ scripts/drive/postgres/thread.lua
@@ -1,5 +1,5 @@
 local effil = require("effil")
-local Request = require("postgres.request")
+local Request = require("drive.postgres.request")
 local luasql = require("luasql.postgres").postgres()
 local threadHandler = require('threadHandler')
